refactor: UserService 리펙토링
